Fixed altref to work
The workaround is like this.  Whenever an altref frame is received, data from the altref
frame and a data size is prepended to the interframe data.

The muxer must then parse this and break the altref frame up.  This is not an ideal workaround
as now non WebM muxers will run into errors in using altref frames with the vp8 encoder.  As quicktime
requires input frames to be matched with output frames, there is really no other way to do this
than to have special knowledge of the creation of altref frames within the muxer.

Change-Id: I41624c53d5e4369b53198d67b19d20f97b48b022