Update patch set 1
Patch Set 1:

Are you certain you want to be this close to the bleeding edge of libvpx?  Leaving the submodules pointing at cayuga (0.9.7-p1) would be fine. The old revision is also perfectly alright as well.

The only reason I bring it up is that I'm certain the repo HEAD gets broken occasionally (as is the case with all projects), and I'm not certain the revision you've selected is a known good rev on all platforms.

Patch-set: 1
Label: Code-Review=0
Label: Verified=0