Rewrite installer, add XiphQT and updater.
Add separate installers for the XiphQT and WebM components, and a
third installer that installs the GoogleSoftwareUpdate, XiphQT, WebM
packages.

Add preinstall scripts for each QuickTime component. Both remove
existing versions of the component being installed. Note that the
preinstall scripts are also run at preupgrade time.

Add GoogleSoftwareUpdate.pkg and contents to provide automatic
updates after WebM QuickTime package installation.

Add postinstall scripts to register each component with the updater.
Note that the postinstall scripts are also run at postupgrade time.

Add a build script for creation of the XiphQT and WebM component
installers. The final installer must be built manually-- this is
because PackageMaker *always* crashes when attempting to build the
package on the command line.

Delete the original installer.

Change-Id: Iadfd80befb61d4431164dbc1da9b5ac49f51355c