Update patch set 1
Patch Set 1: Verified

Omaha config generation script-- needed because making the omaha config by hand for mac installs is a bit on the fiddly side.

Note: this does only the config for the full installation package. With the way things are working I might have to abandon the plan to update WebM and XiphQT individually. I don't want to register 3 configs and worry about 3 download pages, and it looks like that's the way it would have to be for 3 payloads when each has a unique app ID.

Patch-set: 1
Label: Code-Review=0
Label: Verified=+1