generate_omaha_config.sh: correct DMG file name.
Include the version number in the DMG file name, and add a TODO
about how the DMG file name is obtained. The former because of
the change to build_dmg.sh, and the latter because the way it's
being done right now is terribly fragile.

Change the download URL to what will ultimately be used, excluding
ossible version number changes.

Add documentation comment for the generate_config function.

Remove deprecated NeedsAdmin flag from generated configuration data.

Change-Id: I89f16ca81ab49b34f932a03b516166ef91094b08